mcp_todo/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
msal>=1.28.0
python-dotenv>=1.0.0
